pyproject.toml
[project]
name = "graphql-fastapi-mongo"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "strawberry-graphql[fastapi] (>=0.264.0,<0.265.0)",
    "beanie (>=1.29.0,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "uvicorn (>=0.34.1,<0.35.0)"
]

[tool.poetry]
packages = [{include = "backend",from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
